% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_objects.R
\name{Analyze.dataDescription.features.categorical}
\alias{Analyze.dataDescription.features.categorical}
\title{Analyze.dataDescription.features.categorical Object}
\usage{

  Analyze.dataDescription.features.categorical(Analyze.dataDescription.features.categorical.values = NULL,
  count = NULL, values = NULL)
}
\arguments{
\item{Analyze.dataDescription.features.categorical.values}{The \link{Analyze.dataDescription.features.categorical.values} object or list of objects}

\item{count}{Number of categorical values for this feature in the data}

\item{values}{List of all the categories for this feature in the data set}
}
\value{
Analyze.dataDescription.features.categorical object
}
\description{
Analyze.dataDescription.features.categorical Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Description of the categorical values of this feature.
}
\seealso{
Other Analyze functions: \code{\link{Analyze.dataDescription.features.categorical.values}},
  \code{\link{Analyze.dataDescription.features.numeric}},
  \code{\link{Analyze.dataDescription.features.text}},
  \code{\link{Analyze.dataDescription.features}},
  \code{\link{Analyze.dataDescription.outputFeature.numeric}},
  \code{\link{Analyze.dataDescription.outputFeature.text}},
  \code{\link{Analyze.dataDescription.outputFeature}},
  \code{\link{Analyze.dataDescription}},
  \code{\link{Analyze.modelDescription.confusionMatrixRowTotals}},
  \code{\link{Analyze.modelDescription.confusionMatrix}},
  \code{\link{Analyze.modelDescription}},
  \code{\link{Analyze}}
}
