% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_objects.R
\name{Analyze.dataDescription.features.numeric}
\alias{Analyze.dataDescription.features.numeric}
\title{Analyze.dataDescription.features.numeric Object}
\usage{
Analyze.dataDescription.features.numeric(count = NULL, mean = NULL,
  variance = NULL)
}
\arguments{
\item{count}{Number of numeric values for this feature in the data set}

\item{mean}{Mean of the numeric values of this feature in the data set}

\item{variance}{Variance of the numeric values of this feature in the data set}
}
\value{
Analyze.dataDescription.features.numeric object
}
\description{
Analyze.dataDescription.features.numeric Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Description of the numeric values of this feature.
}
\seealso{
Other Analyze functions: \code{\link{Analyze.dataDescription.features.categorical.values}},
  \code{\link{Analyze.dataDescription.features.categorical}},
  \code{\link{Analyze.dataDescription.features.text}},
  \code{\link{Analyze.dataDescription.features}},
  \code{\link{Analyze.dataDescription.outputFeature.numeric}},
  \code{\link{Analyze.dataDescription.outputFeature.text}},
  \code{\link{Analyze.dataDescription.outputFeature}},
  \code{\link{Analyze.dataDescription}},
  \code{\link{Analyze.modelDescription.confusionMatrixRowTotals}},
  \code{\link{Analyze.modelDescription.confusionMatrix}},
  \code{\link{Analyze.modelDescription}},
  \code{\link{Analyze}}
}

