% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_objects.R
\name{Analyze.dataDescription.features.text}
\alias{Analyze.dataDescription.features.text}
\title{Analyze.dataDescription.features.text Object}
\usage{
Analyze.dataDescription.features.text(count = NULL)
}
\arguments{
\item{count}{Number of multiple-word text values for this feature}
}
\value{
Analyze.dataDescription.features.text object
}
\description{
Analyze.dataDescription.features.text Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Description of multiple-word text values of this feature.
}
\seealso{
Other Analyze functions: \code{\link{Analyze.dataDescription.features.categorical.values}},
  \code{\link{Analyze.dataDescription.features.categorical}},
  \code{\link{Analyze.dataDescription.features.numeric}},
  \code{\link{Analyze.dataDescription.features}},
  \code{\link{Analyze.dataDescription.outputFeature.numeric}},
  \code{\link{Analyze.dataDescription.outputFeature.text}},
  \code{\link{Analyze.dataDescription.outputFeature}},
  \code{\link{Analyze.dataDescription}},
  \code{\link{Analyze.modelDescription.confusionMatrixRowTotals}},
  \code{\link{Analyze.modelDescription.confusionMatrix}},
  \code{\link{Analyze.modelDescription}},
  \code{\link{Analyze}}
}

