% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_objects.R
\name{Analyze.dataDescription.outputFeature}
\alias{Analyze.dataDescription.outputFeature}
\title{Analyze.dataDescription.outputFeature Object}
\usage{

  Analyze.dataDescription.outputFeature(Analyze.dataDescription.outputFeature.numeric = NULL,
  Analyze.dataDescription.outputFeature.text = NULL, numeric = NULL,
  text = NULL)
}
\arguments{
\item{Analyze.dataDescription.outputFeature.numeric}{The \link{Analyze.dataDescription.outputFeature.numeric} object or list of objects}

\item{Analyze.dataDescription.outputFeature.text}{The \link{Analyze.dataDescription.outputFeature.text} object or list of objects}

\item{numeric}{Description of the output values in the data set}

\item{text}{Description of the output labels in the data set}
}
\value{
Analyze.dataDescription.outputFeature object
}
\description{
Analyze.dataDescription.outputFeature Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Description of the output value or label.
}
\seealso{
Other Analyze functions: \code{\link{Analyze.dataDescription.features.categorical.values}},
  \code{\link{Analyze.dataDescription.features.categorical}},
  \code{\link{Analyze.dataDescription.features.numeric}},
  \code{\link{Analyze.dataDescription.features.text}},
  \code{\link{Analyze.dataDescription.features}},
  \code{\link{Analyze.dataDescription.outputFeature.numeric}},
  \code{\link{Analyze.dataDescription.outputFeature.text}},
  \code{\link{Analyze.dataDescription}},
  \code{\link{Analyze.modelDescription.confusionMatrixRowTotals}},
  \code{\link{Analyze.modelDescription.confusionMatrix}},
  \code{\link{Analyze.modelDescription}},
  \code{\link{Analyze}}
}

