% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_objects.R
\name{Analyze.dataDescription}
\alias{Analyze.dataDescription}
\title{Analyze.dataDescription Object}
\usage{
Analyze.dataDescription(Analyze.dataDescription.features = NULL,
  Analyze.dataDescription.features.categorical = NULL,
  Analyze.dataDescription.features.categorical.values = NULL,
  Analyze.dataDescription.features.numeric = NULL,
  Analyze.dataDescription.features.text = NULL,
  Analyze.dataDescription.outputFeature = NULL,
  Analyze.dataDescription.outputFeature.numeric = NULL,
  Analyze.dataDescription.outputFeature.text = NULL, features = NULL,
  outputFeature = NULL)
}
\arguments{
\item{Analyze.dataDescription.features}{The \link{Analyze.dataDescription.features} object or list of objects}

\item{Analyze.dataDescription.features.categorical}{The \link{Analyze.dataDescription.features.categorical} object or list of objects}

\item{Analyze.dataDescription.features.categorical.values}{The \link{Analyze.dataDescription.features.categorical.values} object or list of objects}

\item{Analyze.dataDescription.features.numeric}{The \link{Analyze.dataDescription.features.numeric} object or list of objects}

\item{Analyze.dataDescription.features.text}{The \link{Analyze.dataDescription.features.text} object or list of objects}

\item{Analyze.dataDescription.outputFeature}{The \link{Analyze.dataDescription.outputFeature} object or list of objects}

\item{Analyze.dataDescription.outputFeature.numeric}{The \link{Analyze.dataDescription.outputFeature.numeric} object or list of objects}

\item{Analyze.dataDescription.outputFeature.text}{The \link{Analyze.dataDescription.outputFeature.text} object or list of objects}

\item{features}{Description of the input features in the data set}

\item{outputFeature}{Description of the output value or label}
}
\value{
Analyze.dataDescription object
}
\description{
Analyze.dataDescription Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Description of the data the model was trained on.
}
\seealso{
Other Analyze functions: \code{\link{Analyze.dataDescription.features.categorical.values}},
  \code{\link{Analyze.dataDescription.features.categorical}},
  \code{\link{Analyze.dataDescription.features.numeric}},
  \code{\link{Analyze.dataDescription.features.text}},
  \code{\link{Analyze.dataDescription.features}},
  \code{\link{Analyze.dataDescription.outputFeature.numeric}},
  \code{\link{Analyze.dataDescription.outputFeature.text}},
  \code{\link{Analyze.dataDescription.outputFeature}},
  \code{\link{Analyze.modelDescription.confusionMatrixRowTotals}},
  \code{\link{Analyze.modelDescription.confusionMatrix}},
  \code{\link{Analyze.modelDescription}},
  \code{\link{Analyze}}
}

