% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_objects.R
\name{Analyze.modelDescription.confusionMatrixRowTotals}
\alias{Analyze.modelDescription.confusionMatrixRowTotals}
\title{Analyze.modelDescription.confusionMatrixRowTotals Object}
\usage{
Analyze.modelDescription.confusionMatrixRowTotals()
}
\value{
Analyze.modelDescription.confusionMatrixRowTotals object
}
\description{
Analyze.modelDescription.confusionMatrixRowTotals Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of the confusion matrix row totals
}
\seealso{
Other Analyze functions: \code{\link{Analyze.dataDescription.features.categorical.values}},
  \code{\link{Analyze.dataDescription.features.categorical}},
  \code{\link{Analyze.dataDescription.features.numeric}},
  \code{\link{Analyze.dataDescription.features.text}},
  \code{\link{Analyze.dataDescription.features}},
  \code{\link{Analyze.dataDescription.outputFeature.numeric}},
  \code{\link{Analyze.dataDescription.outputFeature.text}},
  \code{\link{Analyze.dataDescription.outputFeature}},
  \code{\link{Analyze.dataDescription}},
  \code{\link{Analyze.modelDescription.confusionMatrix}},
  \code{\link{Analyze.modelDescription}},
  \code{\link{Analyze}}
}

