% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_objects.R
\name{Analyze.modelDescription}
\alias{Analyze.modelDescription}
\title{Analyze.modelDescription Object}
\usage{
Analyze.modelDescription(Analyze.modelDescription.confusionMatrix = NULL,
  Analyze.modelDescription.confusionMatrixRowTotals = NULL,
  confusionMatrix = NULL, confusionMatrixRowTotals = NULL,
  modelinfo = NULL)
}
\arguments{
\item{Analyze.modelDescription.confusionMatrix}{The \link{Analyze.modelDescription.confusionMatrix} object or list of objects}

\item{Analyze.modelDescription.confusionMatrixRowTotals}{The \link{Analyze.modelDescription.confusionMatrixRowTotals} object or list of objects}

\item{confusionMatrix}{An output confusion matrix}

\item{confusionMatrixRowTotals}{A list of the confusion matrix row totals}

\item{modelinfo}{Basic information about the model}
}
\value{
Analyze.modelDescription object
}
\description{
Analyze.modelDescription Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Description of the model.
}
\seealso{
Other Analyze functions: \code{\link{Analyze.dataDescription.features.categorical.values}},
  \code{\link{Analyze.dataDescription.features.categorical}},
  \code{\link{Analyze.dataDescription.features.numeric}},
  \code{\link{Analyze.dataDescription.features.text}},
  \code{\link{Analyze.dataDescription.features}},
  \code{\link{Analyze.dataDescription.outputFeature.numeric}},
  \code{\link{Analyze.dataDescription.outputFeature.text}},
  \code{\link{Analyze.dataDescription.outputFeature}},
  \code{\link{Analyze.dataDescription}},
  \code{\link{Analyze.modelDescription.confusionMatrixRowTotals}},
  \code{\link{Analyze.modelDescription.confusionMatrix}},
  \code{\link{Analyze}}
}
