% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_objects.R
\name{Analyze}
\alias{Analyze}
\title{Prediction API Objects 
Lets you access a cloud hosted machine learning service that makes it easy to build smart apps}
\usage{
Analyze(Analyze.dataDescription = NULL,
  Analyze.dataDescription.features = NULL,
  Analyze.dataDescription.features.categorical = NULL,
  Analyze.dataDescription.features.categorical.values = NULL,
  Analyze.dataDescription.features.numeric = NULL,
  Analyze.dataDescription.features.text = NULL,
  Analyze.dataDescription.outputFeature = NULL,
  Analyze.dataDescription.outputFeature.numeric = NULL,
  Analyze.dataDescription.outputFeature.text = NULL,
  Analyze.modelDescription = NULL,
  Analyze.modelDescription.confusionMatrix = NULL,
  Analyze.modelDescription.confusionMatrixRowTotals = NULL,
  dataDescription = NULL, errors = NULL, id = NULL,
  modelDescription = NULL, selfLink = NULL)
}
\arguments{
\item{Analyze.dataDescription}{The \link{Analyze.dataDescription} object or list of objects}

\item{Analyze.dataDescription.features}{The \link{Analyze.dataDescription.features} object or list of objects}

\item{Analyze.dataDescription.features.categorical}{The \link{Analyze.dataDescription.features.categorical} object or list of objects}

\item{Analyze.dataDescription.features.categorical.values}{The \link{Analyze.dataDescription.features.categorical.values} object or list of objects}

\item{Analyze.dataDescription.features.numeric}{The \link{Analyze.dataDescription.features.numeric} object or list of objects}

\item{Analyze.dataDescription.features.text}{The \link{Analyze.dataDescription.features.text} object or list of objects}

\item{Analyze.dataDescription.outputFeature}{The \link{Analyze.dataDescription.outputFeature} object or list of objects}

\item{Analyze.dataDescription.outputFeature.numeric}{The \link{Analyze.dataDescription.outputFeature.numeric} object or list of objects}

\item{Analyze.dataDescription.outputFeature.text}{The \link{Analyze.dataDescription.outputFeature.text} object or list of objects}

\item{Analyze.modelDescription}{The \link{Analyze.modelDescription} object or list of objects}

\item{Analyze.modelDescription.confusionMatrix}{The \link{Analyze.modelDescription.confusionMatrix} object or list of objects}

\item{Analyze.modelDescription.confusionMatrixRowTotals}{The \link{Analyze.modelDescription.confusionMatrixRowTotals} object or list of objects}

\item{dataDescription}{Description of the data the model was trained on}

\item{errors}{List of errors with the data}

\item{id}{The unique name for the predictive model}

\item{modelDescription}{Description of the model}

\item{selfLink}{A URL to re-request this resource}
}
\value{
Analyze object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:06:39
filename: /Users/mark/dev/R/autoGoogleAPI/googlepredictionv15.auto/R/prediction_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Analyze Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Analyze functions: \code{\link{Analyze.dataDescription.features.categorical.values}},
  \code{\link{Analyze.dataDescription.features.categorical}},
  \code{\link{Analyze.dataDescription.features.numeric}},
  \code{\link{Analyze.dataDescription.features.text}},
  \code{\link{Analyze.dataDescription.features}},
  \code{\link{Analyze.dataDescription.outputFeature.numeric}},
  \code{\link{Analyze.dataDescription.outputFeature.text}},
  \code{\link{Analyze.dataDescription.outputFeature}},
  \code{\link{Analyze.dataDescription}},
  \code{\link{Analyze.modelDescription.confusionMatrixRowTotals}},
  \code{\link{Analyze.modelDescription.confusionMatrix}},
  \code{\link{Analyze.modelDescription}}
}
