% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_objects.R
\name{List}
\alias{List}
\title{List Object}
\usage{
List(items = NULL, nextPageToken = NULL, selfLink = NULL)
}
\arguments{
\item{items}{List of models}

\item{nextPageToken}{Pagination token to fetch the next page, if one exists}

\item{selfLink}{A URL to re-request this resource}
}
\value{
List object
}
\description{
List Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

