% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_functions.R
\name{trainedmodels.analyze}
\alias{trainedmodels.analyze}
\title{Get analysis of the model and the data the model was trained on.}
\usage{
trainedmodels.analyze(id)
}
\arguments{
\item{id}{The unique name for the predictive model}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/prediction
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/prediction)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/prediction/docs/developer-guide}{Google Documentation}
}
