% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_objects.R
\name{Training}
\alias{Training}
\title{Training Object}
\usage{
Training(Training.modelInfo = NULL, Training.trainingInstances = NULL,
  created = NULL, id = NULL, modelInfo = NULL, modelType = NULL,
  selfLink = NULL, storageDataLocation = NULL, storagePMMLLocation = NULL,
  storagePMMLModelLocation = NULL, trainingComplete = NULL,
  trainingInstances = NULL, trainingStatus = NULL, utility = NULL)
}
\arguments{
\item{Training.modelInfo}{The \link{Training.modelInfo} object or list of objects}

\item{Training.trainingInstances}{The \link{Training.trainingInstances} object or list of objects}

\item{created}{Insert time of the model (as a RFC 3339 timestamp)}

\item{id}{The unique name for the predictive model}

\item{modelInfo}{Model metadata}

\item{modelType}{Type of predictive model (classification or regression)}

\item{selfLink}{A URL to re-request this resource}

\item{storageDataLocation}{Google storage location of the training data file}

\item{storagePMMLLocation}{Google storage location of the preprocessing pmml file}

\item{storagePMMLModelLocation}{Google storage location of the pmml model file}

\item{trainingComplete}{Training completion time (as a RFC 3339 timestamp)}

\item{trainingInstances}{Instances to train model on}

\item{trainingStatus}{The current status of the training job}

\item{utility}{A class weighting function, which allows the importance weights for class labels to be specified [Categorical models only]}
}
\value{
Training object
}
\description{
Training Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Training functions: \code{\link{Training.modelInfo}},
  \code{\link{Training.trainingInstances}},
  \code{\link{trainedmodels.insert}}
}

