% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_objects.R
\name{Training.trainingInstances}
\alias{Training.trainingInstances}
\title{Training.trainingInstances Object}
\usage{
Training.trainingInstances()
}
\value{
Training.trainingInstances object
}
\description{
Training.trainingInstances Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Training functions: \code{\link{Training.modelInfo}},
  \code{\link{Training}},
  \code{\link{trainedmodels.insert}}
}

