% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_objects.R
\name{Update}
\alias{Update}
\title{Update Object}
\usage{
Update(csvInstance = NULL, label = NULL, output = NULL)
}
\arguments{
\item{csvInstance}{The input features for this instance}

\item{label}{The class label of this instance}

\item{output}{The generic output value - could be regression value or class label}
}
\value{
Update object
}
\description{
Update Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Update functions: \code{\link{trainedmodels.update}}
}
